+++ lib/src/service/tweets/tweets_service.dart
@@ -8,12 +8,14 @@ import 'package:twitter_api_core/twitter_api_core.dart' as core;
 // Project imports:
 import '../base_service.dart';
 import '../common/includes.dart';
+import '../common/rate_limit.dart';
 import '../filtered_stream_response.dart';
 import '../media/media_field.dart';
 import '../places/place_field.dart';
 import '../polls/poll_field.dart';
 import '../response_field.dart';
 import '../twitter_response.dart';
+import '../twitter_stream_response.dart';
 import '../users/user_data.dart';
 import '../users/user_expansion.dart';
 import '../users/user_field.dart';
@@ -143,7 +145,7 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/manage-tweets/api-reference/delete-tweets-id
-  Future<bool> destroyTweet({required String tweetId});
+  Future<TwitterResponse<bool, void>> destroyTweet({required String tweetId});
 
   /// Causes the user ID identified in the path parameter to Like the target
   /// Tweet.
@@ -180,7 +182,8 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/likes/api-reference/post-users-id-likes
-  Future<bool> createLike({required String userId, required String tweetId});
+  Future<TwitterResponse<bool, void>> createLike(
+      {required String userId, required String tweetId});
 
   /// Allows a user or authenticated user ID to unlike a Tweet.
   ///
@@ -220,7 +223,8 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/likes/api-reference/delete-users-id-likes-tweet_id
-  Future<bool> destroyLike({required String userId, required String tweetId});
+  Future<TwitterResponse<bool, void>> destroyLike(
+      {required String userId, required String tweetId});
 
   /// Causes the user ID identified in the path parameter to Retweet the target
   /// Tweet.
@@ -258,7 +262,8 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/retweets/api-reference/post-users-id-retweets
-  Future<bool> createRetweet({required String userId, required String tweetId});
+  Future<TwitterResponse<bool, void>> createRetweet(
+      {required String userId, required String tweetId});
 
   /// Allows a user or authenticated user ID to remove the Retweet of a Tweet.
   ///
@@ -300,7 +305,7 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/retweets/api-reference/delete-users-id-retweets-tweet_id
-  Future<bool> destroyRetweet(
+  Future<TwitterResponse<bool, void>> destroyRetweet(
       {required String userId, required String tweetId});
 
   /// Allows you to get information about a Tweet’s liking users.
@@ -1444,7 +1449,7 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/bookmarks/api-reference/post-users-id-bookmarks
-  Future<bool> createBookmark(
+  Future<TwitterResponse<bool, void>> createBookmark(
       {required String userId, required String tweetId});
 
   /// Allows a user or authenticated user ID to remove a Bookmark of a Tweet.
@@ -1482,7 +1487,7 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/bookmarks/api-reference/delete-users-id-bookmarks-tweet_id
-  Future<bool> destroyBookmark(
+  Future<TwitterResponse<bool, void>> destroyBookmark(
       {required String userId, required String tweetId});
 
   /// Allows you to get information about a authenticated user’s 800 most recent
@@ -1613,7 +1618,8 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/hide-replies/api-reference/put-tweets-id-hidden
-  Future<bool> createHiddenReply({required String tweetId});
+  Future<TwitterResponse<bool, void>> createHiddenReply(
+      {required String tweetId});
 
   /// Unhides a reply to a Tweet.
   ///
@@ -1646,7 +1652,8 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/hide-replies/api-reference/put-tweets-id-hidden
-  Future<bool> destroyHiddenReply({required String tweetId});
+  Future<TwitterResponse<bool, void>> destroyHiddenReply(
+      {required String tweetId});
 
   /// Returns Tweets mentioning a single user specified by the requested user
   /// ID.
@@ -2239,7 +2246,8 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/volume-streams/api-reference/get-tweets-sample-stream
-  Future<Stream<TwitterResponse<TweetData, void>>> connectVolumeStream({
+  Future<TwitterStreamResponse<TwitterResponse<TweetData, void>>>
+      connectVolumeStream({
     int? backfillMinutes,
     List<TweetExpansion>? expansions,
     List<TweetField>? tweetFields,
@@ -2356,7 +2364,7 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/filtered-stream/api-reference/get-tweets-search-stream
-  Future<Stream<FilteredStreamResponse>> connectFilteredStream({
+  Future<TwitterStreamResponse<FilteredStreamResponse>> connectFilteredStream({
     int? backfillMinutes,
     List<TweetExpansion>? expansions,
     List<TweetField>? tweetFields,
@@ -2458,7 +2466,7 @@ abstract class TweetsService {
   /// ## Reference
   ///
   /// - https://developer.twitter.com/en/docs/twitter-api/tweets/filtered-stream/api-reference/post-tweets-search-stream-rules
-  Future<bool> destroyFilteringRules({
+  Future<TwitterResponse<bool, void>> destroyFilteringRules({
     required List<String> ruleIds,
   });
 }
@@ -2514,7 +2522,9 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> destroyTweet({required String tweetId}) async =>
+  Future<TwitterResponse<bool, void>> destroyTweet({
+    required String tweetId,
+  }) async =>
       super.evaluateResponse(
         await super.delete(
           core.UserContext.oauth2OrOAuth1,
@@ -2523,7 +2533,7 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> createLike({
+  Future<TwitterResponse<bool, void>> createLike({
     required String userId,
     required String tweetId,
   }) async =>
@@ -2536,7 +2546,7 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> destroyLike({
+  Future<TwitterResponse<bool, void>> destroyLike({
     required String userId,
     required String tweetId,
   }) async =>
@@ -2548,7 +2558,7 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> createRetweet({
+  Future<TwitterResponse<bool, void>> createRetweet({
     required String userId,
     required String tweetId,
   }) async =>
@@ -2561,7 +2571,7 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> destroyRetweet({
+  Future<TwitterResponse<bool, void>> destroyRetweet({
     required String userId,
     required String tweetId,
   }) async =>
@@ -2890,7 +2900,7 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> createBookmark({
+  Future<TwitterResponse<bool, void>> createBookmark({
     required String userId,
     required String tweetId,
   }) async =>
@@ -2903,7 +2913,7 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> destroyBookmark({
+  Future<TwitterResponse<bool, void>> destroyBookmark({
     required String userId,
     required String tweetId,
   }) async =>
@@ -2942,7 +2952,8 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> createHiddenReply({required String tweetId}) async =>
+  Future<TwitterResponse<bool, void>> createHiddenReply(
+          {required String tweetId}) async =>
       super.evaluateResponse(
         await super.put(
           core.UserContext.oauth2OrOAuth1,
@@ -2952,7 +2963,8 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<bool> destroyHiddenReply({required String tweetId}) async =>
+  Future<TwitterResponse<bool, void>> destroyHiddenReply(
+          {required String tweetId}) async =>
       super.evaluateResponse(
         await super.put(
           core.UserContext.oauth2OrOAuth1,
@@ -3083,7 +3095,8 @@ class _TweetsService extends BaseService implements TweetsService {
       );
 
   @override
-  Future<Stream<TwitterResponse<TweetData, void>>> connectVolumeStream({
+  Future<TwitterStreamResponse<TwitterResponse<TweetData, void>>>
+      connectVolumeStream({
     int? backfillMinutes,
     List<TweetExpansion>? expansions,
     List<TweetField>? tweetFields,
@@ -3106,18 +3119,24 @@ class _TweetsService extends BaseService implements TweetsService {
       },
     );
 
-    return stream.map(
-      (event) => TwitterResponse(
-        data: TweetData.fromJson(event[ResponseField.data.value]),
-        includes: event.containsKey(ResponseField.includes.value)
-            ? Includes.fromJson(event[ResponseField.includes.value])
-            : null,
+    final headers = super.headerConverter.convert(stream.headers);
+
+    return TwitterStreamResponse(
+      rateLimit: RateLimit.fromJson(headers),
+      stream: stream.body.map(
+        (event) => TwitterResponse(
+          rateLimit: RateLimit.fromJson(headers),
+          data: TweetData.fromJson(event[ResponseField.data.value]),
+          includes: event.containsKey(ResponseField.includes.value)
+              ? Includes.fromJson(event[ResponseField.includes.value])
+              : null,
+        ),
       ),
     );
   }
 
   @override
-  Future<Stream<FilteredStreamResponse>> connectFilteredStream({
+  Future<TwitterStreamResponse<FilteredStreamResponse>> connectFilteredStream({
     int? backfillMinutes,
     List<TweetExpansion>? expansions,
     List<TweetField>? tweetFields,
@@ -3140,15 +3159,21 @@ class _TweetsService extends BaseService implements TweetsService {
       },
     );
 
-    return stream.map(
-      (event) => FilteredStreamResponse(
-        data: TweetData.fromJson(event[ResponseField.data.value]),
-        includes: event.containsKey(ResponseField.includes.value)
-            ? Includes.fromJson(event[ResponseField.includes.value])
-            : null,
-        matchingRules: (event[ResponseField.matchingRules.value] as List)
-            .map((json) => MatchingRule.fromJson(json))
-            .toList(),
+    final headers = super.headerConverter.convert(stream.headers);
+
+    return TwitterStreamResponse(
+      rateLimit: RateLimit.fromJson(headers),
+      stream: stream.body.map(
+        (event) => FilteredStreamResponse(
+          rateLimit: RateLimit.fromJson(headers),
+          data: TweetData.fromJson(event[ResponseField.data.value]),
+          includes: event.containsKey(ResponseField.includes.value)
+              ? Includes.fromJson(event[ResponseField.includes.value])
+              : null,
+          matchingRules: (event[ResponseField.matchingRules.value] as List)
+              .map((json) => MatchingRule.fromJson(json))
+              .toList(),
+        ),
       ),
     );
   }
@@ -3192,7 +3217,7 @@ class _TweetsService extends BaseService implements TweetsService {
           );
 
   @override
-  Future<bool> destroyFilteringRules({
+  Future<TwitterResponse<bool, void>> destroyFilteringRules({
     required List<String> ruleIds,
   }) async =>
       super.evaluateResponse(