+++ test/src/service/tweets/tweets_service_test.dart
@@ -27,6 +27,7 @@ import 'package:twitter_api_v2/src/service/tweets/tweet_poll_param.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweet_reply_param.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweets_service.dart';
 import 'package:twitter_api_v2/src/service/twitter_response.dart';
+import 'package:twitter_api_v2/src/service/twitter_stream_response.dart';
 import 'package:twitter_api_v2/src/service/users/user_data.dart';
 import 'package:twitter_api_v2/src/service/users/user_meta.dart';
 import '../../../mocks/client_context_stubs.dart' as context;
@@ -132,8 +133,8 @@ void main() {
 
       final response = await tweetsService.destroyTweet(tweetId: '1111');
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -174,8 +175,8 @@ void main() {
 
       final response = await tweetsService.destroyTweet(tweetId: '1111');
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -194,8 +195,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -245,8 +246,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -264,8 +265,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -313,8 +314,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -333,8 +334,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -384,8 +385,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -403,8 +404,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -452,8 +453,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -1555,8 +1556,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -1606,8 +1607,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -1625,8 +1626,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -1674,8 +1675,8 @@ void main() {
         tweetId: '1111',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -1774,8 +1775,8 @@ void main() {
 
       final response = await tweetsService.createHiddenReply(tweetId: '0000');
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -1820,8 +1821,8 @@ void main() {
         tweetId: '0000',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -1836,8 +1837,8 @@ void main() {
 
       final response = await tweetsService.destroyHiddenReply(tweetId: '0000');
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -1882,8 +1883,8 @@ void main() {
         tweetId: '0000',
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });
 
@@ -2320,9 +2321,10 @@ void main() {
         backfillMinutes: 5,
       );
 
-      final data = await response.toList();
+      final data = await response.stream.toList();
 
-      expect(response, isA<Stream<TwitterResponse<TweetData, void>>>());
+      expect(response,
+          isA<TwitterStreamResponse<TwitterResponse<TweetData, void>>>());
       expect(data, isA<List<TwitterResponse<TweetData, void>>>());
       expect(data.length, 70);
     });
@@ -2343,7 +2345,9 @@ void main() {
       final data = <TwitterResponse<TweetData, void>>[];
       final errors = <dynamic>[];
 
-      await for (final event in response.handleError(errors.add)) {
+      final stream = response.stream;
+
+      await for (final event in stream.handleError(errors.add)) {
         data.add(event);
       }
 
@@ -2366,9 +2370,9 @@ void main() {
       backfillMinutes: 5,
     );
 
-    final data = await response.toList();
+    final data = await response.stream.toList();
 
-    expect(response, isA<Stream<FilteredStreamResponse>>());
+    expect(response, isA<TwitterStreamResponse<FilteredStreamResponse>>());
     expect(data.first, isA<FilteredStreamResponse>());
     expect(data.first.data, isA<TweetData>());
     expect(data.first.matchingRules, isA<List<MatchingRule>>());
@@ -2428,8 +2432,8 @@ void main() {
         ruleIds: ['XXXX', 'YYYY'],
       );
 
-      expect(response, isA<bool>());
-      expect(response, isTrue);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isTrue);
     });
 
     test('with invalid access token', () async {
@@ -2474,8 +2478,8 @@ void main() {
         ruleIds: ['XXXX', 'YYYY'],
       );
 
-      expect(response, isA<bool>());
-      expect(response, isFalse);
+      expect(response, isA<TwitterResponse<bool, void>>());
+      expect(response.data, isFalse);
     });
   });