+++ CHANGELOG.md
@@ -1,5 +1,64 @@
 # Release Note
 
+## v4.0.0
+
+### New Feature
+
+- Supported the feature for getting the rate limit of each endpoints from response. Now you can get `RateLimit` object from response and it has `limitCount`, `remainingCount` and `resetAt` fields. ([#440](https://github.com/twitter-dart/twitter-api-v2/issues/440))
+
+### Destructive Change
+
+- It was necessary to modify the response structure for endpoints that use `Stream` to include a `RateLimit` object in each endpoint response. The modified methods are following.
+  - `connectVolumeStream`
+  - `connectFilteredStream`
+
+- Also, methods that previously returned the result of processing as a `bool` have been modified to return a `TwitterResponse`. This is to have a `RateLimit` object as above. So, for example, the `createLike` method can also obtain a `RateLimit` object, and the `bool` value of the processing result can be obtained from the TwitterResponse's `data` field.
+
+And now the responses from the modified methods can be used as follows:
+
+```dart
+import 'package:twitter_api_v2/twitter_api_v2.dart' as v2;
+
+Future<void> main() async {
+  final twitter = v2.TwitterApi(bearerToken: 'YOUR_TOKEN_HERE');
+
+  try {
+    final response = await twitter.tweetsService.connectFilteredStream();
+    print(response.rateLimit);
+
+    final stream = response.stream;
+    await for (final event in stream.handleError(print)) {
+      print(event);
+    }
+  } on v2.TwitterException catch (e) {
+    print(e);
+  }
+}
+```
+
+```dart
+import 'package:twitter_api_v2/twitter_api_v2.dart' as v2;
+
+Future<void> main() async {
+  final twitter = v2.TwitterApi(bearerToken: 'YOUR_TOKEN_HERE');
+
+  try {
+    final me = await twitter.usersService.lookupMe();
+    final tweets = await twitter.tweetsService.searchRecent(query: '#ElonMusk');
+
+    final response = await twitter.tweetsService.createLike(
+      userId: me.data.id,
+      tweetId: tweets.data.first.id,
+    );
+
+    print(response.rateLimit);
+    print(response.data); // true or false
+  } on v2.TwitterException catch (e) {
+    print(e);
+  }
+}
+```
+
 ## v3.2.3
 
 - Refactored the internal process for `MediaService`.